// frontend/src/types/babysitter.ts
export type Babysitter = {
    id: number;
    name: string;
    cpf: string;
    dateOfBirth: string;
    address: string;
    phoneNumber: string;
    email: string;
    weeklyHours: number;
    monthlySalary: number;
    bonus: number;
    yearsOfExperience: number;
    maxTravelDistance: number;
};